Clean up fs
